fix: resolve 100+ Kotlin compilation errors across 40+ files
- Add missing imports for IntentResult, IntentType, Trend, TaskComplexity
- Fix Trace -> ReActTrace reference in ConversationManager
- Fix DOT_MATCHES_MULTILINE -> DOT_MATCHES_ALL in OutputSanitizer
- Fix val reassignment in ReasoningChain, VoicePipelineHarness
- Fix type mismatches (Int/Long, Float/Double) in ReasoningModelManager, ModelVersionManager, DifferentialPrivacy
- Fix missing TaskType enum values (INVENTORY_OPTIMIZATION, etc.)
- Fix xor operations on Byte in QuantumReadyLayer
- Fix nullable receiver in SwahiliParser
- Add LAST_WRITE_WINS to ResolutionAction enum
- Fix DialectAdapterFactory to instantiate adapters
- Fix SecurityModule to pass MsaidiziApi to OtpManager
- Fix EncryptedStorage to use getOrCreateKey
- Fix MlDsaProvider/MlKemProvider PQC parameter types
- Fix DeviceTier.HIGH -> Tier.ENHANCED references
- Fix navigation actions in nav_graph.xml
- Fix ModelDownloadManager to inject ModelDownloader
- Fix BootstrapViewModel when exhaustiveness
- Fix ConversationLearningPipeline WordCapture params
- Fix PeerComparison PeerMetrics periodStart
- Fix VoicePipeline isBasicTier initialization
- Fix LocalStsProvider session, onToken, tier issues
- Fix AgentNamingFragment customInput/selectedDisplay scope
- Fix VoiceSetupFragment lifecycleScope import
- Fix CI workflow: remove non-existent jacocoTestReportDebug task
- Add recordTransaction to BusinessAgent
- Add BusinessPatternTracker import to MorningBriefingLoop
- Add kotlin.math.ln/exp imports to FederatedLearningPrivacy

+++ app/src/main/java/com/msaidizi/app/agent/version/ModelVersionManager.kt
@@ -166,7 +166,7 @@ class ModelVersionManager(private val context: Context) {
         val expectedBytes = version.quantizedSizeMb.toLong() * 1024 * 1024
         val actualBytes = file.length()
         val tolerance = expectedBytes * 0.2
-        return actualBytes in (expectedBytes - tolerance)..(expectedBytes + tolerance)
+        return actualBytes.toDouble() in (expectedBytes - tolerance)..(expectedBytes + tolerance)
     }
 
     /**
